Added float variables to the game.

+++ Engine/Face.h
@@ -5,8 +5,8 @@
 class Face
 {
 private:
-	int x = 400;
-	int y = 300;
+	float x = 400.0f;
+	float y = 300.0f;
 	static constexpr int width = 19;
 	static constexpr int height = 19;
 
@@ -14,8 +14,8 @@ public:
 	void BorderTest();
 	void Draw(Graphics& gfx) const;
 	void Update(MainWindow& wnd);
-	int GetX() const;
-	int GetY() const;
+	float GetX() const;
+	float GetY() const;
 	const int GetWidth() const;
 	const int GetHeight() const;
 };
